restructure the utils folder into a "models-utils" and "simple utils"
