+++ src/index.ts
@@ -3,15 +3,14 @@ export * from '@models/token/ERC20/erc20-token-lock';
 export * from '@models/token/ERC721/erc721-collectibles';
 export * from '@models/token/ERC721/erc721-standard';
 export * from '@models/token/ERC1155/erc1155-ownable';
+export * from '@models/token/ERC4626/erc4626';
 export * from '@models/bepro/network-v2';
 export * from '@models/bepro/token/ERC721/bounty-token';
 export * from '@models/staking-contract';
 export * from '@models/votable';
 export * from '@models/bepro/network-registry';
-export * from '@models/token/ERC4626/erc4626';
 
 export * from '@base/model';
-export * from '@interfaces/deployable';
 export * from '@base/ownable';
 export * from '@base/web3-connection';
 export * from '@base/web3-contract';
@@ -25,39 +24,38 @@ export * from '@interfaces/bepro/network-issue';
 export * from '@interfaces/bepro/treasury-info';
 export * from '@interfaces/bepro/oracles-summary';
 export * from '@interfaces/bepro/oracles-resume';
-export * from '@interfaces/staking-product';
-export * from '@interfaces/staking-subscription';
-export * from '@interfaces/swap-router';
-export * from '@interfaces/voter-info';
-export * from '@interfaces/modifiers';
-export * from '@interfaces/block-number-timestamp';
-export * from '@interfaces/web3-connection-options';
 export * from '@interfaces/bepro/bounty';
 export * from '@interfaces/bepro/pull-request';
 export * from '@interfaces/bepro/proposal';
 export * from '@interfaces/bepro/proposal-detail';
 export * from '@interfaces/bepro/benefactor';
+export * from '@interfaces/staking-product';
+export * from '@interfaces/staking-subscription';
+export * from '@interfaces/voter-info';
+export * from '@interfaces/modifiers';
+export * from '@interfaces/block-number-timestamp';
+export * from '@interfaces/web3-connection-options';
 export * from '@interfaces/eip4361';
 export * from '@interfaces/typed-data-v4';
 
-export * from '@events/x-events';
-export * from '@events/network-events';
-export * from '@events/erc4626';
+export * from '@utils/models/benefactor';
+export * from '@utils/models/bounty';
+export * from '@utils/models/delegation';
+export * from '@utils/models/locked-tokens-info';
+export * from '@utils/models/oracles-resume';
+export * from '@utils/models/pool-information';
+export * from '@utils/models/proposal';
+export * from '@utils/models/proposal-detail';
+export * from '@utils/models/pull-request';
+export * from '@utils/models/stake-subscription';
+export * from '@utils/models/staking-product';
+export * from '@utils/models/transaction-handler';
+export * from '@utils/models/voter-info';
+export * from '@utils/models/voting-poll-winner';
 
-export * from '@utils/locked-tokens-info';
-export * from '@utils/numbers';
-export * from '@utils/pool-information';
-export * from '@utils/stake-subscription';
-export * from '@utils/transaction-handler';
-export * from '@utils/voter-info';
-export * from '@utils/voting-poll-winner';
-export * from '@utils/benefactor';
-export * from '@utils/bounty';
-export * from '@utils/oracles-resume';
-export * from '@utils/proposal-detail';
-export * from '@utils/proposal';
-export * from '@utils/pull-request';
+export * from '@utils/constants';
 export * from '@utils/eip4361';
 export * from '@utils/json-rpc-params';
+export * from '@utils/numbers';
 
-export * as Defaults from '@utils/constants';
+
