--- src/interfaces/events/governed-events.ts
@@ -1,3 +1,0 @@
-import {CustomEvent} from "@events/x-events";
-
-export type GovernorTransferredEvent = CustomEvent<{returnValues: {'previousGovernor': string;'newGovernor': string;}}>
