+++ src/models/staking-contract.ts
@@ -5,10 +5,10 @@ import {Deployable} from '@interfaces/deployable';
 import {ERC20} from '@models/token/ERC20/erc20';
 import {ERC721Collectibles} from '@models/token/ERC721/erc721-collectibles';
 import {fromDecimals, toSmartContractDate, toSmartContractDecimals} from '@utils/numbers';
-import stakingProduct from '@utils/staking-product';
+import stakingProduct from '@utils/models/staking-product';
 import {IsOwnable} from '@interfaces/modifiers';
 import {Ownable} from '@base/ownable';
-import stakeSubscription from '@utils/stake-subscription';
+import stakeSubscription from '@utils/models/stake-subscription';
 import {Errors} from '@interfaces/error-enum';
 import {nativeZeroAddress} from '@utils/constants';
 import artifact from "@interfaces/generated/abi/StakingContract";
