+++ src/utils/models/oracles-resume.ts
