+++ src/utils/models/proposal-detail.ts
