+++ src/utils/models/staking-product.ts
