+++ src/utils/models/voter-info.ts
