+++ src/utils/models/voting-poll-winner.ts
